Added all final programs and assests
Uploaded the entire code for the game split into separate pieces. Each piece hosts the code of a certain element of the game, like main mechanics, player class, loading assets, and physics/collisions and so on.
Also uploaded all assets of the game. "Sounds" hosts all the sound effects. "Songs" hosts the normal version of the song and the reversed version. "Images" hosts all the sprites, backgrounds, and icons.